Refactor buttons as new reusable component
